refactor:MemoDialog共通部分まとめる #9
